+++ packages/cashc/src/cashc-cli.ts
@@ -5,13 +5,15 @@ import {
   calculateBytesize,
   countOpcodes,
   exportArtifact,
+  formatArtifact,
   scriptToAsm,
   scriptToBytecode,
 } from '@cashscript/utils';
-import { program } from 'commander';
+import { program, Option } from 'commander';
 import fs from 'fs';
 import path from 'path';
 import { compileFile, version } from './index.js';
+import { MAX_INPUT_BYTESIZE } from './constants.js';
 
 program
   .storeOptionsAsProperties(false)
@@ -23,6 +25,11 @@ program
   .option('-A, --asm', 'Compile the contract to ASM format rather than a full artifact.')
   .option('-c, --opcount', 'Display the number of opcodes in the compiled bytecode.')
   .option('-s, --size', 'Display the size in bytes of the compiled bytecode.')
+  .addOption(
+    new Option('-f, --format <format>', 'Specify the format of the output.')
+      .choices(['json', 'ts'])
+      .default('json'),
+  )
   .helpOption('-?, --help', 'Display help')
   .parse();
 
@@ -48,11 +55,8 @@ function run(): void {
     const opcount = countOpcodes(script);
     const bytesize = calculateBytesize(script);
 
-    if (opcount > 201) {
-      console.warn('Warning: Your contract\'s opcount is over the limit of 201 and will not be accepted by the BCH network');
-    }
-    if (bytesize > 520) {
-      console.warn('Warning: Your contract\'s bytesize is over the limit of 520 and will not be accepted by the BCH network');
+    if (bytesize > MAX_INPUT_BYTESIZE) {
+      console.warn(`Warning: Your contract is ${bytesize} bytes, which is over the limit of ${MAX_INPUT_BYTESIZE} bytes and will not be accepted by the BCH network`);
     }
 
     if (opts.asm) {
@@ -82,10 +86,10 @@ function run(): void {
       if (!fs.existsSync(outputDir)) {
         fs.mkdirSync(outputDir, { recursive: true });
       }
-      exportArtifact(artifact, outputFile);
+      exportArtifact(artifact, outputFile, opts.format);
     } else {
       // Output artifact to STDOUT
-      console.log(JSON.stringify(artifact, null, 2));
+      console.log(formatArtifact(artifact, opts.format));
     }
   } catch (e: any) {
     abort(e.message);